Create machine struct for evaluation a lir statement
