+++ src/lib.rs
@@ -3,12 +3,14 @@
 use std::io::{self, Write};
 
 use thiserror::Error;
+use crate::machine::Machine;
 
 pub mod ast;
 pub mod lir;
 pub mod mir;
 pub mod parser;
 pub mod ty;
+pub mod machine;
 
 pub type LangResult<T> = Result<T, LangError>;
 
@@ -38,6 +40,8 @@ pub fn run_with_env(input: &str, env: &mut LangEnv) -> LangResult<lir::Term> {
     let mir = mir::Term::from_ast(ast)?;
     let _ty = ty::ty_check(&mir)?;
     let lir = lir::Term::from_mir(mir);
-    let res = lir::evaluate(lir, env);
+    let res = Machine {
+        env
+    }.evaluate(lir);
     Ok(res)
 }
